requirements.txt
requests>=2.31.0
acachecontrol>=0.3.5
retry2
aiohttp>=3.9.2
certifi>=2023.7.22
urllib3>=1.26.18
setuptools>=65.5.1
pyjwt==2.8.0
cryptography==42.0.8